requirements.txt
Django==1.6.4
MySQL-python==1.2.5
django-extensions==1.3.3
djangorestframework==2.3.13
djangorestframework-bulk==0.1.1
gunicorn==18.0
pep8==1.4.6
pytz==2013.9
six==1.5.2
wsgiref==0.1.2
